Refactor, folder movement up and down, and also sideways
