+++ lib/actor-system/actor-system.ts
@@ -8,14 +8,14 @@
 import Fiber from '../fiber/fiber'
 import IProcessor from '../fiber/processor'
 import Mailbox from '../mailbox/mailbox'
+import IMaterializer from '../materializer/materializer'
+import IResolver from '../resolver/resolver'
 import Actor, { IActor } from './actor'
 import ActorMessage from './actor-message'
 import ActorProxy from './actor-proxy'
 import IActorSystemConfiguration from './configuration/actor-system-configuration'
 import ActorSystemConfigurationBuilder from './configuration/actor-system-configuration-builder'
-import FunctionActor from './fn/function'
-import IMaterializer from './materializer/materializer'
-import IResolver from './resolver/resolver'
+import FunctionActor from './function'
 import IActorSupervisor from './supervision/actor-supervisor'
 
 export default class ActorSystem implements IProcessor {
